+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+
+## [1.1.0] - 2015-05-11
 ### Added
 - Remove exception if Response in not OK
 - Five more tests to Response/DecryptResponse
@@ -28,6 +30,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - WSCryptDecryptSoapClient class
 - PHPUnit Test Suite
 
-[unreleased]: https://github.com/endelwar/GestPayWS/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/endelwar/GestPayWS/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/endelwar/GestPayWS/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/endelwar/GestPayWS/compare/v0.1.0...v1.0.0
 [0.1.0]: https://github.com/endelwar/GestPayWS/compare/67d07c5c9c4d1873ba9620af25b91e0a53664d80...v0.1.0